Propagate errors between websocket threads
Resolves #453.

Here's a breakdown of the mistake I made:

Supposed the thread processing incoming JSON messages encounters a network
exception and exits. I had thought that the other threads would then be
cancelled and we'd reconnect. In reality, the other threads would
continue running and we would rely on either ping timeout or deadlock
detection to get us out of the loop. Not good!

The lesson here is that you have to `wait` for your threads. That's the
error propagation mechanism between threads.