Open new store connection for each push
This commit introduces a change to open a new connection to the store for each push operation.
Previously, a single connection was opened at the beginning of a Cachix invocation,
leading to issues when reading the SQLite database in immutable mode.

For instance, long-running commands like watch-* would not recognize new valid paths,
as the database would not contain the new entries.

To address this issue, we now open a new database connection whenever the push logic is invoked.