Filter invalid paths when watching the store
It appears that Nix can produce "ghost paths" when a derivation's output
is symlinked to another a path, like in the case of the referenced
libidn derivation.

Even if this is a bug in Nix, this store path now exists (or doesn't)
and we need to filter it out. In the case of the watch-style commands,
we had filtering on the consuming side, but not on the
producing/querying side. This would lead to a crash when computing the
closure.

Fixes #547.