Set the file system encoding to utf8
Filesystem operations in Haskell respect the system locale, which may or
may not be set correctly. A utf8-encoded path on a (supposedly) non-utf8
system will be mangled when read into Haskell.

On unix, we can assume utf8 encoding. A more robust solution would be to
update hnix-store-core and cachix to use OsPath, dodging the whole
encoding question.

Resolves #571.