daemon: load the typical user nix.conf files
When NIX_USER_CONF_FILES is set, nix will only load the confs specified
by it. For our use-case, we want the client to load the same user
nix.confs that it normally does. To make that happen we need to
explicity add the default locations for all the various confs and
replicate the logic that nix uses.