cachix: fix handling of optional `!include` nix.conf statements
Optional includes would throw an error if the file was missing.

The issue was that the check handling missing/required files did not
implement all of the branches correctly.

Instead of first checking the file and then reading it, we now read the
file and handle the errors after the fact. This makes the logic easier
to understand at a glance.